+++ stumpy/aampdist_snippets.py
@@ -86,7 +86,7 @@ def _get_all_aampdist_profiles(
         T = np.pad(T, pad_width, mode="constant", constant_values=np.nan)
 
     n_padded = T.shape[0]
-    D = np.empty(((n_padded // m) - 1, n_padded - m + 1))
+    D = np.empty(((n_padded // m) - 1, n_padded - m + 1), dtype=np.float64)
 
     if s is not None:
         s = min(int(s), m)
@@ -214,11 +214,11 @@ def aampdist_snippets(
 
     snippets = np.empty((k, m))
     snippets_indices = np.empty(k, dtype=np.int64)
-    snippets_profiles = np.empty((k, D.shape[-1]))
-    snippets_fractions = np.empty(k)
-    snippets_areas = np.empty(k)
-    Q = np.full(D.shape[-1], np.inf)
-    indices = np.arange(0, n_padded - m, m)
+    snippets_profiles = np.empty((k, D.shape[-1]), dtype=np.float64)
+    snippets_fractions = np.empty(k, dtype=np.float64)
+    snippets_areas = np.empty(k, dtype=np.float64)
+    Q = np.full(D.shape[-1], np.inf, dtype=np.float64)
+    indices = np.arange(0, n_padded - m, m, dtype=np.int64)
     snippets_regimes_list = []
 
     for i in range(k):
